Add missing DateSerializers for circe
